+++ lib/generators/recording_studio_stripe/install/templates/recording_studio_stripe_initializer.rb
@@ -41,6 +41,7 @@ RecordingStudioStripe.configure do |config|
   #   }
   # }
   # config.limit_reached_path = "/plans"
+  # config.usage_path = "/billing/usage"
   # config.automatic_tax = true
   # config.allow_promotion_codes = true
 end